Continue the development of hhvm/actions under a new licensor
Changes the Facebook Inc. Copyright statement from `2017-present` to `2017-2022`.
The last commit made by a Facebook Inc. (Meta) employee was made on June 6th, 2022.
https://github.com/hhvm/actions/commit/e726916ede120394e3ff29d7476b63ab69c7587f
I created this fork on July 6th, 2024.
I am the copyright holder of all changes made after this date.
The terms of the license remain the same; anyone, Facebook Inc. (Meta) included,
is allowed to use this Software under the terms of the MIT license.